pyproject.toml
[tool.black]
line-length = 99
include = '/project$|\.pyi?$'
exclude = '''
/(
  \.git
  | \.venv
  | \.vscode
)/
'''

[tool.codespell]
ignore-words-list = "astroid,comparision,connexion,doubleclick,Partiton,partiton"
skip = "./.venv./dbt/dbt_packages,.dbt/logs,.dbt/target,pyproject.toml"

[tool.isort]
profile = "black"

[tool.poetry]
name = "dynamic-task-mapping-for-dbt"
version = "0.0.0"
description = ""
authors = ["padraic.slattery@xebia.com"]
packages = []

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
apache-airflow=">=2.7,<2.8"
apache-airflow-providers-airbyte="^3.3"
apache-airflow-providers-cncf-kubernetes="^7.3"
apache-airflow-providers-slack="^7.3"
astronomer-cosmos = "1.0.5"
dbt-airflow-factory=">=0.34"
dbt-bigquery = ">=1.5.0"
openlineage-airflow="*"

[tool.poetry.dev-dependencies]
pre-commit = "^3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
